+++ src/lib/commentPopup.tsx
@@ -10,9 +10,8 @@ export default function CommentPopup(props: {
     showReplies?: boolean;
     comment: CommentType;
     fetchComment?: boolean;
-    openComment?: boolean;
 }) {
-    const { open, setOpen, comment, showReplies, fetchComment, openComment } = props;
+    const { open, setOpen, comment, showReplies, fetchComment } = props;
     const [isExpanded, setIsExpanded] = useState(!!showReplies);
     const isSmallScreen = useIsSmallScreen();
     return (
@@ -36,7 +35,6 @@ export default function CommentPopup(props: {
                 inPopUp
                 setIsExpanded={setIsExpanded}
                 showReplies={showReplies}
-                openComment={openComment}
             />
         </PopUp>
     );